feat/refact: add functionality to filter books
