+++ pruebas/01-reading-list/cmglezpdev/src/layout/AppLayout.tsx
@@ -2,7 +2,6 @@ import { FC, useContext } from 'react';
 import { Link } from 'react-router-dom';
 import { AppContext } from '../context/AppContext';
 
-import { Menu } from '../components';
 import '../styles/AppLayout.css';
 
 interface Props {
@@ -32,10 +31,7 @@ export const AppLayout:FC<Props> = ({ children }) => {
             </header>
         
             <main className='main_container'>
-                <Menu books={location.pathname === '/' ? availableBooks : readingBooks} />
-                <section className='main__content'>
-                    { children }
-                </section>
+                { children }
             </main>
       </>
     )
