+++ pruebas/01-reading-list/cmglezpdev/src/pages/ReadingList.tsx
@@ -1,16 +1,26 @@
-import { useContext } from 'react';
-import { BookItem } from '../components';
+import { useContext, useEffect } from 'react';
+import { BookItem, FilterGenresMenu } from '../components';
 import { AppContext } from '../context/AppContext';
 import { AppLayout } from '../layout';
+import { useFilterBooks } from '../hook';
 
 export const ReadingList = () => {
-  const { readingBooks } = useContext(AppContext);
+  const { readingBooks: books } = useContext(AppContext);
+  const { filterOptions, toggleSelectOption, setOptions, filterBooks } = useFilterBooks([]);
+
+  useEffect(() => {
+      const genres: string[] = 
+          [...new Set(books.map(book => book.genre))]
+          .sort((a, b) => a.localeCompare(b));
+        setOptions(genres);
+  }, [books]);
 
   return (
     <AppLayout>
+       <FilterGenresMenu books={books} filters={filterOptions} toggleSelect={toggleSelectOption}  />
       <section className='library'>
         {
-          readingBooks.map(book => (
+          filterBooks(books).map(book => (
             <BookItem key={book.ISBN} book={book} removeFromReadingButton />
           ))
         }
